requirements.txt
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
atpublic==5.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
entrypoints==0.4
filelock==3.16.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
narwhals==1.7.0
numpy==2.1.1
openai==1.44.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.2
plotly==5.15.0
pluggy==1.5.0
protobuf==3.20.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
Pympler==1.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requirements-parser==0.9.0
rich==13.7.1
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowcli==0.1.1
snowflake-cli-labs==2.8.1
snowflake-connector-python==3.11.0
snowflake-snowpark-python==1.21.1
snowflake.core==0.8.0
sortedcontainers==2.4.0
st-pages==1.0.1
streamlit==1.38.0
streamlit-bookmark==0.1.2
streamlit-feedback==0.1.2
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.0
toolz==0.9.0
tornado==6.4.1
tqdm==4.66.5
trubrics==1.6.2
typer==0.12.3
types-setuptools==74.1.0.20240907
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==4.0.2
win32-setctime==1.1.0
zipp==3.20.1
